--- a/datalog_test.go
+++ b/datalog_test.go
@@ -21,43 +21,61 @@ func TestDataLog(t *testing.T) {
 		{"fox", 64},
 		{"jumps", 81},
 		{"over", 97},
-		{"the", 97},
 		{"lazy", 113},
 		{"dog", 128},
 	}
 
 	var datalog []byte
 
-	t.Run("put", func(t *testing.T) {
+	t.Run("Write", func(t *testing.T) {
 		dr := bytes.NewReader([]byte{})
 		dw := bytes.NewBuffer([]byte{})
-		idx, err := NewIndex(dr)
-		assert.NoError(err)
-		dl := NewDatalog(dr, dw, idx)
+		dl := NewDatalog(dr, dw)
 		for _, tt := range datalogtests {
 			data := []byte(tt.in)
-			expectedScore := MakeScore(data)
-			score, err := dl.Put(data)
+			n, err := dl.Write(data)
 			assert.NoError(err)
-			assert.Equal(expectedScore, score)
+			expectedSize := intSize + scoreSize + len(data)
+			assert.Equal(expectedSize, n)
 			assert.Len(dw.Bytes(), tt.size)
 		}
 		datalog = make([]byte, dw.Len())
 		copy(datalog, dw.Bytes())
 	})
 
-	t.Run("get", func(t *testing.T) {
+	t.Run("ReadAt", func(t *testing.T) {
 		dr := bytes.NewReader(datalog)
 		dw := bytes.NewBuffer([]byte{})
-		idx, err := NewIndex(dr)
-		assert.NoError(err)
-		dl := NewDatalog(dr, dw, idx)
-		for _, tt := range datalogtests {
+		dl := NewDatalog(dr, dw)
+		for i, tt := range datalogtests {
 			expectedData := []byte(tt.in)
-			score := MakeScore(expectedData)
-			data, err := dl.Get(score)
+			expectedSize := intSize + scoreSize + len(expectedData)
+			data := make([]byte, expectedSize)
+			off := 0
+			if i > 0 {
+				off = datalogtests[i-1].size
+			}
+			n, err := dl.ReadAt(data, int64(off))
 			assert.NoError(err)
-			assert.Equal(expectedData, data)
+			assert.Equal(expectedSize, n)
+			_, deserialized := dl.Deserialize(data)
+			assert.Equal(expectedData, deserialized)
+		}
+	})
+
+	t.Run("Serialize/Deserialize", func(t *testing.T) {
+		dr := bytes.NewReader([]byte{})
+		dw := bytes.NewBuffer([]byte{})
+		dl := NewDatalog(dr, dw)
+		for _, tt := range datalogtests {
+			score := Score{}
+			data := []byte(tt.in)
+			serialized := dl.Serialize(score, data)
+			assert.NotEqual(data, serialized)
+			assert.Greater(len(serialized), len(data))
+			score2, deserialized := dl.Deserialize(serialized)
+			assert.Equal(data, deserialized)
+			assert.Equal(score, score2)
 		}
 	})
 }