--- a/index_test.go
+++ b/index_test.go
@@ -1,7 +1,6 @@
 package dieci
 
 import (
-	"bytes"
 	"os"
 	"testing"
 
@@ -17,30 +16,28 @@ func TestIndex(t *testing.T) {
 		addr Addr
 		ok   bool
 	}{
-		{"the", Addr{pos: 0, size: 15}, true},
-		{"quick", Addr{pos: 15, size: 17}, true},
-		{"brown", Addr{pos: 32, size: 17}, true},
-		{"fox", Addr{pos: 49, size: 15}, true},
-		{"jumps", Addr{pos: 64, size: 17}, true},
-		{"over", Addr{pos: 81, size: 16}, true},
+		{"the", Addr{pos: 0, size: 3}, true},
+		{"quick", Addr{pos: 3, size: 5}, true},
+		{"brown", Addr{pos: 8, size: 5}, true},
+		{"fox", Addr{pos: 13, size: 3}, true},
+		{"jumps", Addr{pos: 16, size: 5}, true},
+		{"over", Addr{pos: 21, size: 4}, true},
 		{"missing", Addr{pos: 0, size: 0}, false},
-		{"the", Addr{pos: 0, size: 15}, true},
-		{"lazy", Addr{pos: 97, size: 16}, true},
-		{"dog", Addr{pos: 113, size: 15}, true},
+		{"the", Addr{pos: 0, size: 3}, true},
+		{"lazy", Addr{pos: 25, size: 4}, true},
+		{"dog", Addr{pos: 29, size: 3}, true},
 	}
 
-	t.Run("open empty", func(t *testing.T) {
-		reader := bytes.NewReader([]byte{})
-		idx, err := NewIndex(reader)
-		assert.NoError(err)
+	var index Index
+
+	t.Run("NewIndex", func(t *testing.T) {
+		idx := NewIndex()
 		assert.Len(idx.cache, 0)
 		assert.Equal(0, idx.cur)
 	})
 
-	t.Run("write", func(t *testing.T) {
-		reader := bytes.NewReader([]byte{})
-		idx, err := NewIndex(reader)
-		assert.NoError(err)
+	t.Run("Put", func(t *testing.T) {
+		idx := NewIndex()
 		for _, tt := range idxtests {
 			if !tt.ok {
 				continue
@@ -49,29 +46,19 @@ func TestIndex(t *testing.T) {
 			size := len(data)
 			score := MakeScore(data)
 			before := idx.Len()
-			idx.Write(score, size)
+			idx.Put(score, size)
 			assert.GreaterOrEqual(idx.Len(), before)
 			before = idx.Len()
-			idx.Write(score, 0)
+			idx.Put(score, 0)
 			assert.Equal(idx.Len(), before, "Should ignore same update")
 		}
+		index = *idx
 	})
 
-	t.Run("open", func(t *testing.T) {
-		reader := readerFromDatalog()
-		idx, err := NewIndex(reader)
-		assert.NoError(err)
-		assert.Len(idx.cache, 8)
-		assert.Equal(int(reader.Size()), idx.cur)
-	})
-
-	t.Run("read", func(t *testing.T) {
-		reader := readerFromDatalog()
-		idx, err := NewIndex(reader)
-		assert.NoError(err)
+	t.Run("Get", func(t *testing.T) {
 		for _, tt := range idxtests {
 			score := MakeScore([]byte(tt.in))
-			addr, ok := idx.Read(score)
+			addr, ok := index.Get(score)
 			if tt.ok {
 				assert.True(ok, "Should indicate that score exists")
 			} else {
@@ -82,8 +69,8 @@ func TestIndex(t *testing.T) {
 	})
 }
 
-// BenchmarkOpenIndex for iterative improvement of open
-func BenchmarkOpenIndex(b *testing.B) {
+// BenchmarkIndexLoad for iterative improvement of index load
+func BenchmarkIndexLoad(b *testing.B) {
 	for n := 0; n < b.N; n++ {
 		b.StopTimer()
 		f, err := os.Open("testdata/words.data")
@@ -91,23 +78,11 @@ func BenchmarkOpenIndex(b *testing.B) {
 			b.Fatal(err)
 		}
 		b.StartTimer()
-		_, err = NewIndex(f)
+		idx := NewIndex()
+		err = idx.Load(f)
 		if err != nil {
 			b.Fatal(err)
 		}
 		f.Close()
 	}
-}
-
-// FIXME! - this is a hack, I'm obviously need something better here
-func readerFromDatalog() *bytes.Reader {
-	dr := bytes.NewReader([]byte{})
-	dw := bytes.NewBuffer([]byte{})
-	idx, _ := NewIndex(dr)
-	dl := NewDatalog(dr, dw, idx)
-	for _, word := range []string{"the", "quick", "brown", "fox", "jumps", "over", "lazy", "dog"} {
-		data := []byte(word)
-		dl.Put(data)
-	}
-	return bytes.NewReader(dw.Bytes())
 }