Refactoring.

Make datalog into `Writer` and `ReaderAt`
Move index out of datalog into dieci and make it more map-like