@eiri 2020-05-04
```
$ benchcmp before.bench after.bench 
benchmark                old ns/op     new ns/op     delta
BenchmarkOpenIndex-4     70940032      71095718      +0.22%
BenchmarkOpen-4          61108828      59047369      -3.37%
BenchmarkWrite-4         12182         12169         -0.11%
BenchmarkRead-4          1171          1118          -4.53%

benchmark                old allocs     new allocs     delta
BenchmarkOpenIndex-4     9694           9690           -0.04%
BenchmarkOpen-4          9696           9685           -0.11%
BenchmarkWrite-4         1              1              +0.00%
BenchmarkRead-4          2              1              -50.00%

benchmark                old bytes     new bytes     delta
BenchmarkOpenIndex-4     31408163      31403708      -0.01%
BenchmarkOpen-4          31406655      31400755      -0.02%
BenchmarkWrite-4         1245          1245          +0.00%
BenchmarkRead-4          48            32            -33.33%
```